requirements.txt
discord.py
git+https://github.com/Rapptz/discord-ext-menus
git+https://github.com/Stormtorch002/jishaku
requests
Pillow
colour
akinator.py
owoify-py
asyncpg
prettytable
pytz
emoji
chess
svglib
